requirements.txt
# Excel to PowerPoint Generator v6.0
# Web Framework
fastapi
uvicorn[standard]
python-multipart

# PowerPoint Generation
python-pptx

# Excel Reading
openpyxl

# Windows COM Automation (for chart capture)
pywin32; sys_platform == 'win32'

# Image Processing
Pillow

# Data Validation
pydantic>=2.0

# HTTP Client (for testing)
httpx
